requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Telegram Bot
aiogram==3.2.0
apscheduler==3.10.4

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
aiohttp==3.9.1
httpx==0.25.1

# Utilities
pytz==2023.3.post1
tzlocal==5.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0
